Fix unhandled `None` value in project IDs when listing all datasets (#1413)
When fetching the project_ids, there are times where the project id may
be None, even if that's not supposed to be the case. This causes the
program to break. This patch adds some defensive checks to make sure
that even if somehow we get None for the project_id, it will be ignored
and handled gracefully.